aspeed-boot: allow compressed images
Handle compressed images in the aspeed-images.json by uncompressing
them to a temporary file (and then cleaning it up when done).

An uncompressed image for many platforms is larger than Github allows
without using git-lfs, since they need to be padded to the SPI NOR
size, but since there are a lot of zeros they compress fine and are
accepted.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>